perf: updated lazy-load hook as more global
