+++ src/components/hooks/use-handle-lazy-load.tsx
@@ -1,10 +1,10 @@
-import React, { useEffect, useState } from 'react'
-import { OurPlatforms, WhatOurClientsSay, TradeTypes } from '../../pages/home/_lazy-load'
-import TRADE_DUMMY from 'images/common/trade-type-dummy.png'
-import PLATFORM_DUMMY from 'images/common/platforms-dummy.png'
-import WCS_DUMMY from 'images/common/wcs-dummy.png'
-
-const UseHandleLazyLoad = () => {
+import React, { ReactNode, useEffect, useState } from 'react'
+type templateProps = {
+    lazytemplates?: ReactNode
+    targetId?: string
+    options?: object
+}
+const UseHandleLazyLoad = ({ lazytemplates, targetId, options }: templateProps) => {
     const [isVisible, setIsVisible] = useState(false)
     const handleIntersect = (entries, observer) => {
         const [entry] = entries
@@ -24,52 +24,12 @@ const UseHandleLazyLoad = () => {
     }, [])
 
     const createObserver = () => {
-        const target = document.querySelector('#market-fold')
-        const options = {
-            root: null,
-            rootMargin: '0px',
-            threshold: 0.5,
-        }
-
+        const target = document.querySelector(targetId)
         const observer = new IntersectionObserver(handleIntersect, options)
         observer.observe(target)
     }
 
-    return (
-        isVisible && (
-            <>
-                <TradeTypes
-                    fallback={
-                        <div>
-                            <img
-                                src={TRADE_DUMMY}
-                                style={{ marginLeft: '331px' }}
-                                alt="trade-type-dummy"
-                            />
-                        </div>
-                    }
-                />
-                <OurPlatforms
-                    fallback={
-                        <div>
-                            <img
-                                src={PLATFORM_DUMMY}
-                                style={{ marginLeft: '331px' }}
-                                alt="platform-dummy"
-                            />
-                        </div>
-                    }
-                />
-                <WhatOurClientsSay
-                    fallback={
-                        <div>
-                            <img src={WCS_DUMMY} style={{ marginLeft: '331px' }} alt="wcs-dummy" />
-                        </div>
-                    }
-                />
-            </>
-        )
-    )
+    return isVisible && <>{lazytemplates}</>
 }
 
 export default UseHandleLazyLoad
