+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
-import React from 'react'
-import { Signup } from './home/_lazy-load'
+import React, { ReactNode } from 'react'
+import { Signup, OurPlatforms, WhatOurClientsSay, TradeTypes } from './home/_lazy-load'
 import MarketsFold from './home/_markets_fold'
 import Hero from './home/_hero'
 import UseHandleLazyLoad from 'components/hooks/use-handle-lazy-load'
@@ -8,10 +8,52 @@ import { SEO } from 'components/containers'
 import Layout from 'components/layout/layout'
 import { localize, WithIntl } from 'components/localization'
 import { Appearances } from 'components/custom/signup'
+import TRADE_DUMMY from 'images/common/trade-type-dummy.png'
+import PLATFORM_DUMMY from 'images/common/platforms-dummy.png'
+import WCS_DUMMY from 'images/common/wcs-dummy.png'
 //const MarketsFold = React.lazy(() => import('./home/_markets_fold'));
 const Home = () => {
     /* redirect livechat for en to open live chat popup */
     useOpenLiveChat()
+    const target = '#market-fold'
+    const options = {
+        root: null,
+        rootMargin: '0px',
+        threshold: 0.5,
+    }
+    const templates: ReactNode = (
+        <>
+            <TradeTypes
+                fallback={
+                    <div>
+                        <img
+                            src={TRADE_DUMMY}
+                            style={{ marginLeft: '331px' }}
+                            alt="trade-type-dummy"
+                        />
+                    </div>
+                }
+            />
+            <OurPlatforms
+                fallback={
+                    <div>
+                        <img
+                            src={PLATFORM_DUMMY}
+                            style={{ marginLeft: '331px' }}
+                            alt="platform-dummy"
+                        />
+                    </div>
+                }
+            />
+            <WhatOurClientsSay
+                fallback={
+                    <div>
+                        <img src={WCS_DUMMY} style={{ marginLeft: '331px' }} alt="wcs-dummy" />
+                    </div>
+                }
+            />
+        </>
+    )
     return (
         <Layout type="transparent" margin_top="0">
             <SEO
@@ -25,7 +67,7 @@ const Home = () => {
             />
             <Hero />
             <MarketsFold />
-            <UseHandleLazyLoad />
+            <UseHandleLazyLoad lazytemplates={templates} targetId={target} options={options} />
             <Signup appearance={Appearances.public} />
         </Layout>
     )
